requirements.txt
# SOC Lateral Movement Detection - Python Dependencies
# Install with: pip install -r requirements.txt

requests>=2.28.0
python-dateutil>=2.8.0
